Add Project state abstraction.
